requirements.txt
flask>=3.0.0
pillow>=10.2.0
requests>=2.31.0
numpy>=1.26.2
werkzeug>=3.0.1
opencv-python>=4.8.1.78
scikit-image>=0.23.0
pytesseract>=0.3.10
python-dotenv>=1.0.0
